requirements.txt
numpy==1.21.2
pandas==1.3.3
matplotlib==3.4.3
keras==2.6.0
scikit-learn==0.24.2
tensorflow==2.6.0
Pillow==8.2.0
opencv-python==4.5.3.56
tflite==2.6.0
scipy==1.7.3

# # Web framework (if applicable)
# flask==2.1.1

# # Database library (if applicable)
# sqlalchemy==1.4.25

# # Testing framework and tools (if applicable)
# pytest==6.2.5
# coverage==6.3.2
